services/aws-agent/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
langgraph==0.4.8
langchain==0.3.25
langchain-core==0.3.63
langchain-openai==0.3.19
langchain-google-genai==2.1.5
langchain-ollama==0.3.3
pydantic==2.11.5
boto3==1.38.36
botocore==1.38.36
python-dotenv==1.1.0
httpx==0.28.1
typing-extensions==4.14.0
